+++ changelog.md
@@ -1,6 +1,7 @@
 ## Changes in version 0.6.4.0
 
- * Add `Eq1`, `Ord1`, and `Show1` instances for `Array` and `SmallArray`.
+ * Add `Eq1`, `Ord1`, `Show1`, and `Read1` instances for `Array` and
+   `SmallArray`.
 
  * Improve the test suite. This includes having property tests for
    typeclasses from `base` such as `Eq`, `Ord`, `Functor`, `Applicative`,